+++ back-end/src/database/migrations/1603842731593-CreateTransactions.ts
@@ -16,7 +16,7 @@ export default class CreateTransactions1603842731593
                     },
                     {
                         name: 'item_quantity',
-                        type: 'int',
+                        type: 'decimal',
                     },
                     {
                         name: 'type',
